refactor: upgraded to new csproj format and targeted netstandard 1.0 and 2.0
