--- src/ThirdDrawer.UnitTests/packages.config
@@ -1,5 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<packages>
-  <package id="NUnit" version="2.6.3" targetFramework="net40" />
-  <package id="Shouldly" version="1.1.1.1" targetFramework="net40" />
-</packages>
