+++ src/ThirdDrawer/Extensions/StringExtensionMethods/StringExtensions.cs
@@ -2,8 +2,6 @@
 using System.Collections.Generic;
 using System.Globalization;
 using System.Linq;
-using System.Security.Cryptography;
-using System.Text;
 using System.Text.RegularExpressions;
 
 namespace ThirdDrawer.Extensions.StringExtensionMethods
@@ -28,26 +26,6 @@ namespace ThirdDrawer.Extensions.StringExtensionMethods
             }
 
             return new string(characters.ToArray());
-        }
-
-        public static string HashWithSalt(this string stringToHash, string salt)
-        {
-            var encoding = new UTF8Encoding();
-            var saltBytes = encoding.GetBytes(salt);
-            var hashBytes = encoding.GetBytes(stringToHash);
-
-            for (var i = 0; i < 50; i++)
-            {
-                var saltedBytesToHash = hashBytes.ToList();
-                saltedBytesToHash.AddRange(saltBytes);
-
-                using (var hash = new SHA512Managed())
-                {
-                    hashBytes = hash.ComputeHash(saltedBytesToHash.ToArray());
-                }
-            }
-
-            return Convert.ToBase64String(hashBytes);
         }
 
         public static string CoalesceIfWhiteSpace(this string s, string other)
