+++ src/ThirdDrawer/Extensions/TypeExtensionMethods/TypeExtensions.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using System.Linq;
+using System.Reflection;
 using ThirdDrawer.Extensions.CollectionExtensionMethods;
 
 namespace ThirdDrawer.Extensions.TypeExtensionMethods
@@ -8,15 +9,16 @@ namespace ThirdDrawer.Extensions.TypeExtensionMethods
     {
         public static bool IsInstantiable(this Type type)
         {
-            if (type.IsInterface) return false;
-            if (type.IsAbstract) return false;
-            if (type.ContainsGenericParameters) return false;
+            var typeInfo = type.GetTypeInfo();
+            if (typeInfo.IsInterface) return false;
+            if (typeInfo.IsAbstract) return false;
+            if (typeInfo.ContainsGenericParameters) return false;
             return true;
         }
 
         public static bool IsAssignableTo<TTarget>(this Type type)
         {
-            return typeof (TTarget).IsAssignableFrom(type);
+            return typeof (TTarget).GetTypeInfo().IsAssignableFrom(type.GetTypeInfo());
         }
 
         /// <summary>
@@ -32,14 +34,14 @@ namespace ThirdDrawer.Extensions.TypeExtensionMethods
 
         public static bool IsClosedTypeOf(this Type type, Type openGenericType)
         {
-            if (!openGenericType.IsGenericType) throw new ArgumentException("It's a bit difficult to have a closed type of a non-open-generic type", "openGenericType");
+            if (!openGenericType.GetTypeInfo().IsGenericType) throw new ArgumentException("It's a bit difficult to have a closed type of a non-open-generic type", nameof(openGenericType));
 
-            var interfaces = type.GetInterfaces();
-            var baseTypes = new[] {type}.DepthFirst(t => t.BaseType == null ? new Type[0] : new[] {t.BaseType});
+            var interfaces = type.GetTypeInfo().ImplementedInterfaces;
+            var baseTypes = new[] {type}.DepthFirst(t => t.GetTypeInfo().BaseType == null ? new Type[0] : new[] {t.GetTypeInfo().BaseType});
             var typeAndAllThatThatEntails = new[] {type}.Union(interfaces).Union(baseTypes).ToArray();
-            var genericTypes = typeAndAllThatThatEntails.Where(i => i.IsGenericType);
-            var closedGenericTypes = genericTypes.Where(i => !i.IsGenericTypeDefinition);
-            var assignableGenericTypes = closedGenericTypes.Where(i => openGenericType.IsAssignableFrom(i.GetGenericTypeDefinition()));
+            var genericTypes = typeAndAllThatThatEntails.Where(i => i.GetTypeInfo().IsGenericType);
+            var closedGenericTypes = genericTypes.Where(i => !i.GetTypeInfo().IsGenericTypeDefinition);
+            var assignableGenericTypes = closedGenericTypes.Where(i => openGenericType.GetTypeInfo().IsAssignableFrom(i.GetGenericTypeDefinition().GetTypeInfo()));
 
             return assignableGenericTypes.Any();
         }
@@ -51,7 +53,7 @@ namespace ThirdDrawer.Extensions.TypeExtensionMethods
 
         public static Type[] GetGenericInterfacesClosing(this Type type, Type genericInterface)
         {
-            var genericInterfaces = type.GetInterfaces()
+            var genericInterfaces = type.GetTypeInfo().ImplementedInterfaces
                 .Where(i => i.IsClosedTypeOf(genericInterface))
                 .ToArray();
             return genericInterfaces;
