--- src/packages/repositories.config
@@ -1,4 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<repositories>
-  <repository path="..\ThirdDrawer.UnitTests\packages.config" />
-</repositories>
